+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- VSCode: switch opencode api management to SDK
+
+
 ## [1.4.5] - 2026-01-08
 
 - Chat: added support for model variants (thinking effort).